objanim: ObjAnim_GetCurrentEventCountdown return type u16->int (retail is int) — +6 player fns
Declared u16 but retail treats the return as int: callers doing (...)==0 emitted a spurious
clrlwi r0,r3,16; cmplwi where retail emits cmpwi r3,0. Retyping to int (the definition is a
direct lhz field read, byte-neutral) drops the narrow across all 9 player.c call sites:
playerState39/playerStateIdle/playerStateOnLadder ->100.000, playerState25 98.41->99.34,
playerState1D +0.29, playerStateMoving +0.13. Zero regressions (drearthwarrior calls via its
own cast typedef, unaffected; the definition stays 100).

+++ include/main/objanim.h
@@ -35,7 +35,7 @@ int Object_ObjAnimAdvanceMove(int objAnimHandle,f32 moveStepScale,f32 deltaTime,
                               ObjAnimEventList *events);
 int Object_ObjAnimSetMoveProgress(ObjAnimComponent *objAnim,f32 moveProgress);
 int Object_ObjAnimSetMove(int objAnimHandle,int moveId,f32 moveProgress,int moveControlFlags);
-u16 ObjAnim_GetCurrentEventCountdown(ObjAnimComponent *objAnim);
+int ObjAnim_GetCurrentEventCountdown(ObjAnimComponent *objAnim);
 void ObjAnim_WriteStateWord(ObjAnimComponent *objAnim,int stateIndex,short wordIndex,int value);
 void ObjAnim_SetCurrentEventStepFrames(ObjAnimComponent *objAnim,u32 frameCount);
 int ObjAnim_SampleRootCurvePhase(ObjAnimComponent *objAnim,f32 distance,float *phaseOut);
